tinyint added to ad

+++ mm.sql
@@ -45,7 +45,7 @@ create table ad(
         adId integer primary key,
         title varchar2(25) not null,
         runtime integer not null,
-        in_house integer not null,
+        in_house tinyint not null,
         company varchar2(30),
         profit decimal(6,2) not null
 );
